users/serializers.py
from typing import Any, Dict
from rest_framework.serializers import ModelSerializer
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
from .models import User

class UserSerializer(ModelSerializer):
    class Meta:
        model = User
        fields = ["user_id", "username", "email", "phone"]

class UserSignInSerializer(TokenObtainPairSerializer):
    def validate(self, attrs: Dict[str, Any]) -> Dict[str, str]:
        print("XX", attrs)
        return super().validate(attrs)
